// src/app/models/GetJobTimeCards/JobTimeCard.ts
export class JobTimeCard {
    StartTime: string;
    EndTime: string;
    EmpCode: string;
    Grade: string;
    Operation: string;
    OperationType: string;
    OperationCode: string;
    OperationDesc: string;
    GradeCode: string;
    GradeType: string;
    TotalHrs: string;
    JobNo: string;
    Comments: string;
    DiffInMins: string;
    DiffInHrs: string;
    totalWorkedHrs: string;
    ActivityType: string;
    ActivitySubType: string;
}